Split out window into its own module
