+++ examples/render/Makefile
@@ -41,7 +41,7 @@ packagingFlags = $(ENV_VARS) --excludes "$(packagingExcludes)"
 # Set build vars
 compileFlags += -I $(vendorDir)/vulkan/include -I $(vendorDir)/glfw/include -I $(vendorDir)/glm \
 				-I $(vendorDir)/tinyobjloader -I $(vendorDir)/freetype/include
-linkFlags += -l render -l utils
+linkFlags += -l render -l window -l utils
 
 .PHONY: all render-assets
